Lift DetailValueRow out of the backend detail screen
The title-and-value row was declared inside BackendDetailView, which was
fine while that screen was its only caller. The settings cache section
now renders one too, so the type sat inside a domain that no longer owns
it. It moves to Primitives/Rows beside Row, Header and RowSummary, the
row primitives it belongs with, and gains the preview every file in that
folder carries.
